Enable cksum/crc32/tsort/tty/unlink/uuencode toybox applets
Audited the remaining disabled applets (link, unlink, cksum, tty,
logger, expand, fold, tsort, uudecode, uuencode, cal, split, strings)
for LLP64 pointer-width safety, continuing the which/readlink/stat/
touch/id/xargs batch's own discipline.

Found a real bug: tsort.c's own bsearch()-argument-adjustment trick
round-tripped a char** through unsigned long ("do the usual LP64 trick
to MAKE IT SHUT UP", the comment's own words) -- silently truncates on
Windows LLP64. Changed to uintptr_t; recorded in
shell/toybox/PATCHES.md's existing "Windows LLP64 Pointer-Width Fixes"
section.

Found a second, previously-undocumented registration gap: being in
CRT_TOYBOX_SOURCES and shell/toybox/crt/generated/newtoys.h isn't
enough on its own -- shell/toybox/src/android/linux/generated/
config.h's USE_x(...)/CFG_x macro pair (a checked-in Kconfig-style
snapshot) gates whether a newtoys.h entry actually compiles into
toy_list[] at all. All six of these already had a CFG_x 0/empty USE_x
pair sitting there, so flipping both was enough. Confirmed via a
direct "toybox: Unknown command" failure before the fix.

Left disabled, for real reasons rather than oversight:
- expand/logger/fold/uudecode/cal/split/strings all use GLOBALS(), and
  their state struct is missing from generated/flags.h's union
  entirely -- needs toybox's own mkflags generation pipeline, not a
  hand-edit like config.h above.
- link is LLP64-safe but linkat() (and thus link()) is an
  unconditional ENOTSUP stub on every host today, a real PAL gap.

New regression coverage: crt_mksh_rootfs_llp64_batch_runs in
shell/CMakeLists.txt. All 93 tests pass via a genuine cmake --fresh
reconfigure plus full rebuild.

+++ shell/toybox/src/toys/posix/tsort.c
@@ -50,7 +50,13 @@ static void do_tsort(int fd, char *name)
        otop,     // out at start of this loop over pair[]
        otop2,    // out at start of previous loop over pair[]
        ii, jj, kk;
-  unsigned long sigh;
+  /* CRT: uintptr_t, not upstream's "unsigned long" -- see
+   * shell/toybox/PATCHES.md's "Windows LLP64 Pointer-Width Fixes" entry.
+   * Windows x86_64/aarch64 are LLP64 (long stays 32-bit, pointers are
+   * 64-bit); round-tripping a pointer through unsigned long here silently
+   * truncated it before this fix, corrupting the very address bsearch()
+   * below reads pair data through. */
+  uintptr_t sigh;
 
   // bsearch()'s first argument is the element to search for,
   // and sbse() compares the second element of each pair, so to find
@@ -59,7 +65,7 @@ static void do_tsort(int fd, char *name)
   // typecast to (void *), so do the usual LP64 trick to MAKE IT SHUT UP.
   // (The search function adds 1 to each argument so we never access
   // memory outside the pair.)
-  sigh = ((unsigned long)keep)-sizeof(*keep);
+  sigh = ((uintptr_t)keep)-sizeof(*keep);
 
   // Count input entries in data block read from fd
   if (!(ss = readfd(fd, 0, &plen))) return;
